requirements.txt
aiohttp==3.9.3
evdev==1.7.0
python-socketio==5.11.1
